+++ app/EventTournamentParticipant.php
@@ -45,7 +45,8 @@ class EventTournamentParticipant extends Model
         self::created(function($model){
             if (
                 (!isset($model->event_tournament_team_id) || trim($model->event_tournament_team_id) == '') &&
-                (!$model->pug && $model->event_tournament_team_id == null)
+                (!$model->pug && $model->event_tournament_team_id == null) &&
+                $model->format != 'list'
             ) {
                 $challonge = new Challonge(env('CHALLONGE_API_KEY'));
                 $tournament = $challonge->getTournament($model->eventTournament->challonge_tournament_id);
@@ -59,7 +60,7 @@ class EventTournamentParticipant extends Model
             }
         });
         self::deleting(function($model){
-            if (!$model->pug && $model->event_tournament_team_id == null) {
+            if (!$model->pug && $model->event_tournament_team_id == null && $model->format != 'list') {
                 $challonge = new Challonge(env('CHALLONGE_API_KEY'));
                 $participant = $challonge->getParticipant($model->eventTournament->challonge_tournament_id, $model->challonge_participant_id);
                 if (!$response = $participant->delete()) {
